+++ src/main/java/com/prgrms/kream/domain/style/dto/response/GetFeedResponse.java
@@ -7,7 +7,7 @@ public record GetFeedResponse(
 		Long id,
 		Long authorId,
 		String content,
-		Integer likes,
+		Long likes,
 		List<String> images,
 		LocalDateTime createdAt,
 		LocalDateTime updatedAt
